PVP-229 Fix two crashes when previewing tasks; Separate different utilities into representative files
