+++ app/src/main/java/com/pvp/app/service/TaskServiceImpl.kt
@@ -8,10 +8,10 @@ import com.pvp.app.api.ExperienceService
 import com.pvp.app.api.PointService
 import com.pvp.app.api.TaskService
 import com.pvp.app.api.UserService
-import com.pvp.app.common.JSON
-import com.pvp.app.common.resetTime
-import com.pvp.app.common.toJsonElement
-import com.pvp.app.common.toPrimitivesMap
+import com.pvp.app.common.util.DateUtil.resetTime
+import com.pvp.app.common.util.JsonUtil.JSON
+import com.pvp.app.common.util.JsonUtil.toJsonElement
+import com.pvp.app.common.util.JsonUtil.toPrimitivesMap
 import com.pvp.app.model.MealTask
 import com.pvp.app.model.Points
 import com.pvp.app.model.SportActivity
